+++ packages/vips/src/worker.ts
@@ -19,7 +19,6 @@ import {
 	convertImageFormat,
 	compressImage,
 	resizeImage,
-	batchResizeImage,
 	rotateImage,
 	hasTransparency,
 } from './index';
@@ -32,7 +31,6 @@ const api = {
 	convertImageFormat,
 	compressImage,
 	resizeImage,
-	batchResizeImage,
 	rotateImage,
 	hasTransparency,
 };